+++ lib/src/widgets/tasks/random_tasks/random_tasks.dart
@@ -3,28 +3,49 @@ import 'dart:async';
 import 'package:backstreets_widgets/widgets.dart';
 import 'package:flutter/material.dart';
 
+/// An inherited version of [RandomTasks].
+///
+/// Instances of [RandomTasksProvider] will be used when calling
+/// [RandomTasks.maybeOf()] and [RandomTasks.of()].
+class RandomTasksProvider extends InheritedWidget {
+  /// Create an instance.
+  const RandomTasksProvider({
+    required this.pause,
+    required this.resume,
+    required super.child,
+    super.key,
+  });
+
+  /// The function that will pause the [RandomTasks.tasks].
+  final VoidCallback pause;
+
+  /// The function that will resume the [RandomTasks.tasks].
+  final VoidCallback resume;
+
+  /// Whether this widget should notify listeners.
+  @override
+  bool updateShouldNotify(covariant final RandomTasksProvider oldWidget) =>
+      pause != oldWidget.pause || resume != oldWidget.resume;
+}
+
 /// A widget which runs random [tasks].
 ///
 /// This widget is useful for random events, such as playing background sounds,
 /// spawning game objects, or moving non-player characters.
 ///
-/// Tasks can be paused by calling [InheritedRandomTasks.pause], and resumed
-/// with [InheritedRandomTasks.resume]. You can obtain an [InheritedRandomTasks]
+/// Tasks can be paused by calling [RandomTasksProvider.pause], and resumed
+/// with [RandomTasksProvider.resume]. You can obtain an [RandomTasksProvider]
 /// instance with the [maybeOf] or [of] static methods.
 class RandomTasks extends StatefulWidget {
   /// Create an instance.
-  const RandomTasks({
-    required this.tasks,
-    required this.child,
-    super.key,
-  });
+  const RandomTasks({required this.tasks, required this.child, super.key});
 
   /// Possibly provide an instance.
-  static InheritedRandomTasks? maybeOf(final BuildContext context) =>
-      context.dependOnInheritedWidgetOfExactType<InheritedRandomTasks>();
+  static RandomTasksProvider? maybeOf(final BuildContext context) =>
+      context.dependOnInheritedWidgetOfExactType<RandomTasksProvider>();
 
   /// Provide an instance.
-  static InheritedRandomTasks of(final BuildContext context) =>
+  static RandomTasksProvider of(final BuildContext context) =>
       maybeOf(context)!;
 
   /// The tasks to use.
@@ -75,11 +96,8 @@ class RandomTasksState extends State<RandomTasks> {
 
   /// Build a widget.
   @override
-  Widget build(final BuildContext context) => InheritedRandomTasks(
-        pause: pause,
-        resume: resume,
-        child: widget.child,
-      );
+  Widget build(final BuildContext context) =>
+      RandomTasksProvider(pause: pause, resume: resume, child: widget.child);
 
   /// Pause the tasks.
   void pause() => _paused = true;